Rename apps/App.py to follow snake case convention
